+++ NEWS.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - Require Matrix 1.2-14 or higher
 - AddModuleScore no longer densifies sparse-matrices
 - Various visualization fixes and improvements
+- Default value for latent.vars in FindMarkers/FindAllMarkers changed to NULL.
 
 ## [2.3.0] - 2018-03-22
 ### Added